move: Move modules on separate directory
