+++ src/components/pages/LandingPage.js
@@ -1,6 +1,6 @@
 import React from "react"
 import { SECTION_NAMES } from "../vars"
-import RandomRobotGif from "./RandomRobotGif"
+import RandomRobotGif from "../pagePartials/RandomRobotGif"
 
 class LandingPage extends React.Component {
     pageName = SECTION_NAMES.landingPage
